@@ -5,8 +5,9 @@
   "dependencies": {
     "express": "^4.13.3",
     "hbs": "^3.1.1",
-    "react": "0.13.3",
-    "react-router": "^0.13.3"
+    "react": "^0.14.0",
+    "react-dom": "^0.14.0",
+    "react-router": "^1.0.0-rc3"
   },
   "devDependencies": {
     "babelify": "^6.3.0",
